Refactored Models to allow log in by all members
After evaluating the Slack API, it appears that using Slack as a frontend for QR code delivery
will take significantly longer. This is as the chat.postMessage requires an URL for an
attachment. It will take significantly longer to execute as we will need to interact with
multiple resource (RTM, Web) and will be a complex interaction. For now, this API with a
SPA frontend will become the primary method of usage and hence the refactor of the models.
The Admin model will now be just a view