Implemented basic decorator to validateJWTs; will need to write Bearer Middleware
This commit add a tokenRequired decorator as to provide general JWT validation as needed across the entire application. Note that currently there is a non-compliant use of the Authorization header the test currently does not base64 encode the value as required. This will need to be remedied by a custom middleware which is infuriatingly required because werzkeug only support Basic and Digest Authorization headers, and not terribly effectively either it seems (or that webargs is just not good at parsing headers)